% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsim.R
\docType{data}
\name{dsim}
\alias{dsim}
\title{Simulated fisheries and model fits}
\format{A data frame with 914657 rows and 27 columns. The important ones:
\describe{
  \item{stock_id}{Part of a unique stock ID. Must also ID with \code{sigmaC} and \code{sigmaR} and {iter} (iteration).}
  \item{b_bmsy_est}{Estimated (usually median) B/Bmsy.}
  \item{method_id}{Method used.}
}}
\usage{
dsim
}
\description{
Simulated fisheries and model fits
}
\keyword{datasets}

