% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Make superensemble}
\usage{
make(ntree = 1000, cores = 2, ens_formula = log(bbmsy_true_mean) ~ CMSY +
  COMSIR + Costello + SSCOM + spec_freq_0.05 + spec_freq_0.2)
}
\arguments{
\item{ntree}{Number of trees. Passed to \code{\link[randomForest]{randomForest}}.}

\item{cores}{Number of cores to use}

\item{ens_formula}{Specify the formula used in random forest model}
}
\value{
A list with two elements: \code{model} contains a model from the
  package \pkg{randomForest} and \code{data} contains the data used to
  fit the model.
}
\description{
Fit a superensemble model with a simulated dataset. Note that this will
take a few minutes to run.
}
\examples{
\dontrun{
  x <- make()
  randomForest::partialPlot(x$model, x.var = "CMSY", pred.data = x$data)
}
}

